practica integradora

+++ src/services/products.services.js
@@ -1,9 +1,14 @@
 import { productResponseDto } from "../dto/product-response.dto.js"
+import customErrors from "../errors/customErrors.js"
 import error from "../errors/customErrors.js"
 import productsRepository from "../persistences/mongo/repositories/products.repository.js"
 
-const create = async (data) => {
-    const product = await productsRepository.create(data)
+const create = async (data, user) => {
+    let productData = data
+    if(user.role === "premium"){
+        productData = {...data, owner: user._id}
+    }
+    const product = await productsRepository.create(productData)
     return product
 }
 
@@ -25,7 +30,11 @@ const update = async (id, data) => {
     return product
 }
 
-const deleteOne = async (id) => {
+const deleteOne = async (id, user) => {
+    const productData = await productsRepository.getById(id)
+    if(user.role === "premium" && productData.owner !== user._id) {
+        throw customErrors.unauthorizedError("User not authorized")
+    }
     const product = await productsRepository.deleteOne(id)
     if (!product) throw error.notFoundError(`Product id ${id} not found`)
     return product
